updates, tokens now have an _index attribute
because they are essentially strings, list.index(token) fails if
multiple tokens have the same text. I could of course make genuine
objects like before, but list.index anyway is a search method.
On creation of the tree the index is set.

Also add cut_right() and split_right(), these will be needed when
we are retokenizing a modified part of a document.