add regex.to_string()
this method checks whether a regex is in fact a simple string
without special characters, and returns the converted string
in that case.

this can be used to optimize the case where a lexicon has only one
pattern rule, and no dynamic actions etc, and str.find can be used
for lexing.