add leaf_lexicon attribute to Lexicon
if all possible targets, including the default target, pop out of the
current lexicon, the tree builder will never create child contexts for
a context created for this lexicon. This allows for some optimisations
that can be added to token iteration, etc., because we then know
beforehand that such a context will only contain tokens.