experimental leaf context
this allows for some optimisations because a context created by a
lexicon that is a leaf_lexicon will never have child contexts.

So the last pushed lexicon will create a leaf Context (_LeafContext)
when the lexicon has the leaf_lexicon attribute set to True.

But it must be thought out if it is really never possible that when
a tree is partly updated, a context could contain child contexts;
that's why the asserts in the tree builder.

For example, when pushing multiple lexicons at once, a lexicon that
is not the last could be a lexicon with the leaf_lexicon attribute
set to True. We must really be sure that such a context could not
become the current while updating, and than later a leaf context is
erroneously joined with a context with the same ancestry but not
being a leaf context.

So when ever any of the asserts fires, we know we can't use this
branch :-)