fix treebuilder resume when bygroup() does not use the first part of a match
This happens e.g. when using bygroup(String) with pattern r' (")', i.e.
an ignored space before. The position of the first token is then not
the position where parsing can be started. Another case was bygroup(skip,
String) on a pattern like r'( )(")', which caused the same problem.

Now, we just keep the lexer we created to check the part before the
change. This way we don't restart parsing again with a new lexer, which
fixes the strings that only occur after a space in troff requests.