add using() and action.DelegateAction()
This new dynamic action uses a lexicon to lex the matched text.
All tokens are yielded as one group, but without the tree structure. So
although it works flawlessly, it leaves the desire to implement using()
so that it *retains* the tree structure.

But for that to become true, we need more changes in the infrastructure.
For example, we should somehow know when updating a tree that an underlying
lexicon/context needs to be checked, because the match could have changed.
Smart logic should be implemented to check whether reparsing is then needed.