make restarting tree builds more robust; fix spurious issues
in seldom cases, empty contexts were left in the tree. Restarting a build in
build_new_tree appeared to be the problem, the temporary tree can contain
empty contexts at the beginning or end, but finding a new start position in
the temp tree choked on that.

The fix is two-fold: in a temp tree we don't want to start at the beginning
of it, then just ditch it. And we can't rely on Node.parent_index(), so we
need to use alternative functions that use list.index(), which is slower but
not a problem when finding a start position.

In the process, get_prepared_lexer has been made simpler and more robust.